datastorage: fix ap_array_unlink_entry always returns NULL
For an unknown reason the entry that is returned in ap_array_unlink_entry
was always set to NULL. This commit changes the function so it returns the
ap that should be unlinked. This is needed to free the corresponding
memory that is allocated by the ap struct.

Thanks a lot to ptpt52 for reporting this issue.

Signed-off-by: Nick Hainke <vincent@systemli.org>